code fix

+++ src/daft-core/src/series/ops/sign.rs
@@ -9,9 +9,9 @@ impl Series {
         use DataType::*;
         match self.data_type() {
             UInt8 => Ok(self.u8().unwrap().sign_unsigned()?.into_series()),
-            UInt16 => Ok(self.u8().unwrap().sign_unsigned()?.into_series()),
-            UInt32 => Ok(self.u8().unwrap().sign_unsigned()?.into_series()),
-            UInt64 => Ok(self.u8().unwrap().sign_unsigned()?.into_series()),
+            UInt16 => Ok(self.u16().unwrap().sign_unsigned()?.into_series()),
+            UInt32 => Ok(self.u32().unwrap().sign_unsigned()?.into_series()),
+            UInt64 => Ok(self.u64().unwrap().sign_unsigned()?.into_series()),
             Int8 => Ok(self.i8().unwrap().sign()?.into_series()),
             Int16 => Ok(self.i16().unwrap().sign()?.into_series()),
             Int32 => Ok(self.i32().unwrap().sign()?.into_series()),
